refactored Related to place helper function in another file
